PASE score calculator

Calculates PASE score from raw questionnaire data.

Usage

pase_calc(ds, adjust_work = FALSE)

Arguments

ds

data set

adjust_work

flag to set whether to include 10b type 1. Default is TRUE.

Value

data.frame

Details

Labelling should be as defined by the questionnaire. 02-06 should start with 0:3, 02a-06b should start with 1:4.

Regarding work scoring

The score calculation manual available for the PASE questionnaire, all types of work should be included. According to the article by Washburn RA. et al (1999) sitting work is not included in the item 10 score. This differentiation is added with the option to set adjust_work to exclude item 10b category 1 work (set TRUE).

Regarding output

Output includes sub scores as well as sums, but also to columns assessing data quality and completeness. If any field has not been filled, score_incompletes will return TRUE. If all measures are missing score_missings is TRUE. If adjust_work==TRUE, 10b has to be filled, or score_incompletes will be set TRUE.

Examples

summary(pase_calc(stRoke::pase)[,13])
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     0.0    75.0   128.6   138.9   195.9   574.3