PASE score calculator

Calculates PASE score from raw questionnaire data.

Usage

pase_calc(ds, adjust_work = FALSE, consider.missing = c("Not available"))

Arguments

ds

data set

adjust_work

flag to set whether to include 10b type 1.

consider.missing

character vector of values considered missing. Default is TRUE.

Value

data.frame

Details

Labelling should be as defined by the questionnaire. 02-06 should start with 0:3, 02a-06b should start with 1:4.

Regarding work scoring

The score calculation manual available for the PASE questionnaire, all types of work should be included. According to the article by Washburn RA. et al (1999) sitting work is not included in the item 10 score. This differentiation is added with the option to set adjust_work to exclude item 10b category 1 work (set TRUE).

Regarding output

Output includes sub scores as well as sums, but also to columns assessing data quality and completeness. If any field has not been filled, score_incompletes will return TRUE. If all measures are missing score_missings is TRUE. If adjust_work==TRUE, 10b has to be filled, or score_incompletes will be set TRUE.

Examples

summary(pase_calc(stRoke::pase)[,13])
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     0.0    75.0   128.6   138.9   195.9   574.3 
str(pase_calc(stRoke::pase))
#> 'data.frame':	642 obs. of  15 variables:
#>  $ pase_score_02         : num  85.8 5 25.8 25.8 2.2 15 0 51.4 51.4 15 ...
#>  $ pase_score_03         : num  53.97 0 6.72 2.31 2.31 ...
#>  $ pase_score_04         : num  98.67 0 0 0 2.53 ...
#>  $ pase_score_05         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pase_score_06         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pase_score_07         : num  25 25 25 25 25 25 25 25 25 0 ...
#>  $ pase_score_08         : num  25 0 25 25 25 0 25 25 25 0 ...
#>  $ pase_score_09a        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ pase_score_09b        : num  36 0 0 36 0 0 0 0 36 0 ...
#>  $ pase_score_09c        : num  20 0 20 20 0 0 0 0 20 0 ...
#>  $ pase_score_09d        : num  0 35 35 35 0 0 0 0 0 0 ...
#>  $ pase_score_10         : num  12 NA NA 105 9 NA NA 240 NA NA ...
#>  $ pase_score_sum        : num  356 65 138 274 66 ...
#>  $ pase_score_missings   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ pase_score_incompletes: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...